Changed all private instance variables that are "id" to long (from int). Implemented methods for CellarDAO. Added setter and getter for cellarId in Cellar.java

+++ src/com/_500bottles/object/cellar/CellarItem.java
@@ -19,7 +19,7 @@ public class CellarItem
 		return cellarItemId;
 	}
 
-	public void setCellarItemId(int cellarItemId)
+	public void setCellarItemId(long cellarItemId)
 	{
 		this.cellarItemId = cellarItemId;
 	}
